+++ packages/apps/graph-client/src/graphql/fields/non-fungible-chain-account.graph.ts
@@ -16,7 +16,7 @@ export const ALL_NON_FUNGIBLE_CHAIN_ACCOUNT_FIELDS: DocumentNode = gql`
     accountName
     nonFungibleTokenBalances {
       balance
-      id
+      tokenId
     }
 
     # transactions {}