+++ packages/apps/graph/src/graph/subscription/events.ts
@@ -64,32 +64,25 @@ async function* iteratorFn(
   parametersFilter?: string | null,
   minimumDepth?: number | null,
 ): AsyncGenerator<Event[] | undefined, void, unknown> {
-  const eventResult = await getLastEvents(
-    qualifiedEventName,
-    undefined,
-    chainId,
-    parametersFilter,
-    minimumDepth,
-  );
+  let lastEventId;
 
-  let lastEvent;
+  lastEventId = await getLatestEventId();
 
-  if (!nullishOrEmpty(eventResult)) {
-    lastEvent = eventResult[0];
+  if (!nullishOrEmpty(lastEventId)) {
     yield [];
   }
 
   while (!context.req.socket.destroyed) {
     const newEvents = await getLastEvents(
       qualifiedEventName,
-      lastEvent?.id,
+      lastEventId ?? undefined,
       chainId,
       parametersFilter,
       minimumDepth,
     );
 
     if (!nullishOrEmpty(newEvents)) {
-      lastEvent = newEvents[0];
+      lastEventId = newEvents[0].id;
       yield newEvents;
     }
 
@@ -123,8 +116,8 @@ async function getLastEvents(
     where: {
       ...extendedFilter.where,
       qualifiedName: eventName,
-      transaction: {
-        NOT: [],
+      requestKey: {
+        not: 'cb',
       },
       ...(chainId && {
         chainId: parseInt(chainId),
@@ -159,4 +152,17 @@ async function getLastEvents(
   }
 
   return eventsToReturn.sort((a, b) => b.id - a.id);
+}
+
+async function getLatestEventId() {
+  try {
+    const lastEventId = await prismaClient.event.aggregate({
+      _max: {
+        id: true,
+      },
+    });
+    return lastEventId._max.id;
+  } catch (error) {
+    return null;
+  }
 }