+++ packages/apps/graph/src/index.ts
@@ -9,9 +9,8 @@ moduleAlias.addAliases({
   '@devnet': `${__dirname}/devnet`,
 });
 
-import { runSystemsCheck } from '@services/systems-check';
+import { SystemCheckError, runSystemsCheck } from '@services/systems-check';
 import { dotenv } from '@utils/dotenv';
-import { NetworkConfig } from '@utils/network';
 import { useServer } from 'graphql-ws/lib/use/ws';
 import { createYoga } from 'graphql-yoga';
 import 'json-bigint-patch';
@@ -32,8 +31,6 @@ const schema = builder.toSchema();
 
 const plugins = [extensionsPlugin()];
 
-export const networkConfig = NetworkConfig.create(dotenv.NETWORK_HOST);
-
 if (dotenv.COMPLEXITY_EXPOSED) {
   plugins.push(complexityPlugin(schema));
 }
@@ -48,7 +45,6 @@ const yogaApp = createYoga({
   context: async () => {
     return {
       extensions: {},
-      networkId: (await networkConfig).networkId,
     };
   },
 });
@@ -99,7 +95,7 @@ httpServer.on('connection', (socket) => {
   httpServer.once('close', () => sockets.delete(socket));
 });
 
-runSystemsCheck(networkConfig)
+runSystemsCheck()
   .then(() => {
     httpServer.listen(dotenv.PORT, () => {
       console.info(
@@ -107,7 +103,17 @@ runSystemsCheck(networkConfig)
       );
     });
   })
-  .catch(() => {
-    console.log('\nSystem checks failed. Unable to start the graph server.\n');
+  .catch((error) => {
+    if (error instanceof SystemCheckError) {
+      console.log(
+        '\nSystem checks failed. Unable to start the graph server.\n',
+      );
+    } else {
+      console.log(
+        '\nAn unexpected error occurred. Unable to start the graph server.\n',
+      );
+      console.error(error);
+    }
+
     process.exit(1);
   });