+++ packages/libs/react-ui/src/components/Link/Link.stories.tsx
@@ -174,6 +174,54 @@ export const IconOnly: LinkStory = {
   },
 };
 
+export const StartVisualLoading: LinkStory = {
+  args: {
+    variant: 'primary',
+    startVisual: <MonoChevronRight />,
+    children: 'Hello world',
+    isLoading: true,
+  },
+  render: (props: ILinkProps) => {
+    return <Link {...props}>{props.children}</Link>;
+  },
+};
+
+export const EndVisualLoading: LinkStory = {
+  args: {
+    variant: 'primary',
+    endVisual: <MonoChevronRight />,
+    children: 'Hello world',
+    isLoading: true,
+  },
+  render: (props: ILinkProps) => {
+    return <Link {...props}>{props.children}</Link>;
+  },
+};
+
+export const IconOnlyLoadingWithLabel: LinkStory = {
+  args: {
+    variant: 'primary',
+    children: <MonoChevronRight />,
+    isLoading: true,
+    loadingLabel: 'Loading...',
+  },
+  render: (props: ILinkProps) => {
+    return <Link {...props}>{props.children}</Link>;
+  },
+};
+
+export const IconOnlyLoading: LinkStory = {
+  args: {
+    variant: 'primary',
+    children: <MonoChevronRight />,
+    isLoading: true,
+    loadingLabel: '',
+  },
+  render: (props: ILinkProps) => {
+    return <Link {...props}>{props.children}</Link>;
+  },
+};
+
 export const AllVariants: StoryFn<ILinkProps> = ({
   variant,
   ...props